refractor: TemplateCreatorPage -> move event listeners to seperate .ts file
